Fix SID parse issue

+++ SwitchPriv/SwitchPriv/Library/Helpers.cs
@@ -144,7 +144,7 @@ namespace SwitchPriv.Library
             stringSidBuilder.AppendFormat("{0}-{1}", Marshal.ReadByte(pSid), nAuthority);
 
             for (int idx = 0; idx < nSubAuthorityCount; idx++)
-                stringSidBuilder.AppendFormat("-{0}", Marshal.ReadInt32(pSid, 8 + (idx * 4)));
+                stringSidBuilder.AppendFormat("-{0}", (uint)Marshal.ReadInt32(pSid, 8 + (idx * 4)));
 
             stringSid = stringSidBuilder.ToString();
             bSuccess = NativeMethods.LookupAccountSid(
